requirements.txt
Django==4.2.2
Celery==5.3.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
